requirements.txt
flask>=3.0.0
google-generativeai>=0.7.0
groq>=0.9.0
openai>=1.0.0
requests>=2.31.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
python-dotenv>=1.0.0
pygame>=2.5.0
